fix: restore correct ssn validation
